feat(1846): add backend logic to backup provisioner callback
